Add personal debugging output for the prove pass.
The standard debug output appears to reach stderr/stdout NOT
in order of execution. This made it difficult for me to rely
on it for debugging. I suspect this has something to do with
the way in which the logged debug warnings bubble up through
the blocks/node/fns to finally be logged and output.

My debug output, added in this commit, instead uses good-old
fmt.print to emit messages as prove executes. This allows me to
see exactly how a function progresses through the machinery.
I've used the debug flag value of -1 for my additions. This
allows me to use my debug output while silencing the standard
debug output.

Hopefully this approach will allow me to cleanly rebase this
branch onto master without any issues. The alternative was to
comment out the standard debug info. I worry that this wouldnt
rebase well onto an upstream change to that same debug info.

Change-Id: I55a18d3af5dd3a28719c9bd6be9de884175e68e6